--- app/cdap/components/WranglerGrid/AddTransformation/ColumnTable/RadioInput/__test__/RadioInput.test.tsx
@@ -1,39 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-import { fireEvent, render, screen } from '@testing-library/react';
-import React from 'react';
-import RadioInput from 'components/WranglerGrid/AddTransformation/ColumnTable/RadioInput/index';
-
-describe('should test radioInput ', () => {
-  it('Should render the radio component and click the radio', () => {
-    render(
-      <RadioInput
-        selectedColumns={[
-          { name: 'Abhilash', label: 'string', type: [''] },
-          { name: 'Divami', label: 'string', type: [''] },
-        ]}
-        onSingleSelection={jest.fn()}
-        columnDetail={{ name: 'string', label: 'string', type: [''] }}
-        columnIndex={0}
-      />
-    );
-
-    const radioInputElement = screen.getByTestId(/radio-input-0/i);
-    fireEvent.click(radioInputElement);
-    expect(radioInputElement).toBeInTheDocument();
-  });
-});
