+++ app/cdap/components/WranglerGrid/AddTransformationPanel/IconStore/tickIcon.tsx
